Make `Preferences` plain data
Similar to updating the message rates, if we fail to take a lock on a
mutex, we likely have bigger problems. The readers and writers do not
have panic conditions, so something would have to be going wrong higher
up the stack. Remove the use of `Arc` in the handshake and just mutate
the preferences